requirements.txt
python-telegram-bot
sqlalchemy
pandas